Adjust the code structure
