+++ agent/src/ebpf/user/load.c
@@ -33,7 +33,7 @@
 #include <bcc/linux/btf.h>
 #include <bcc/libbpf.h>
 #include "load.h"
-#include "btf_vmlinux.h"
+#include "btf_core.h"
 #include "../kernel/include/bpf_base.h"
 
 extern struct btf_ext *btf_ext__new(const uint8_t * data, uint32_t size);
@@ -78,6 +78,11 @@ void resume_stderr(int fd)
 		return;
 	dup2(fd, STDERR_FILENO);
 	close(fd);
+}
+
+static inline bool str_is_empty(const char *s)
+{
+	return !s || !s[0];
 }
 
 int df_prog_load(enum bpf_prog_type prog_type, const char *name,
@@ -452,360 +457,6 @@ static enum bpf_prog_type get_prog_type(struct sec_desc *desc)
 	return prog_type;
 }
 
-#include "hashmap.h"
-#include "relo_core.h"
-extern const char *btf__name_by_offset(const struct btf *btf, uint32_t offset);
-static inline bool str_is_empty(const char *s)
-{
-	return !s || !s[0];
-}
-
-static bool prog_contains_insn(struct ebpf_prog *prog, size_t insn_idx)
-{
-	return insn_idx >= prog->sec_insn_off &&
-	    insn_idx < prog->sec_insn_off + prog->sec_insn_cnt;
-}
-#if 0
-#define BPF_CORE_SPEC_MAX_LEN 64
-struct bpf_core_accessor {
-	__u32 type_id;		
-	__u32 idx;		
-	const char *name;
-};
-
-struct bpf_core_cand {
-	const struct btf *btf;
-	__u32 id;
-};
-
-struct bpf_core_cand_list {
-	struct bpf_core_cand *cands;
-	int len;
-};
-
-struct bpf_core_spec {
-	const struct btf *btf;
-	struct bpf_core_accessor spec[BPF_CORE_SPEC_MAX_LEN];
-	__u32 root_type_id;
-	enum bpf_core_relo_kind relo_kind;
-	int len;
-	int raw_spec[BPF_CORE_SPEC_MAX_LEN];
-	int raw_len;
-	__u32 bit_offset;
-};
-
-extern const struct btf_type *btf__type_by_id(const struct btf *btf,
-					      __u32 type_id);
-extern const char *btf_kind_str(const struct btf_type *t);
-extern __s64 btf__resolve_size(const struct btf *btf, __u32 type_id);
-extern int bpf_core_parse_spec(const char *prog_name, const struct btf *btf,
-			const struct bpf_core_relo *relo,
-			struct bpf_core_spec *spec);
-extern int bpf_core_format_spec(char *buf, size_t buf_sz, const struct bpf_core_spec *spec);
-#endif
-
-extern const struct btf_type *btf__type_by_id(const struct btf *btf,
-                                              __u32 type_id);
-extern const char *btf_kind_str(const struct btf_type *t);
-extern void bpf_core_free_cands(struct bpf_core_cand_list *cands);
-static bool core_relo_is_type_based(enum bpf_core_relo_kind kind)
-{
-	switch (kind) {
-	case BPF_CORE_TYPE_ID_LOCAL:
-	case BPF_CORE_TYPE_ID_TARGET:
-	case BPF_CORE_TYPE_EXISTS:
-	case BPF_CORE_TYPE_MATCHES:
-	case BPF_CORE_TYPE_SIZE:
-		return true;
-	default:
-		return false;
-	}
-}
-
-static bool core_relo_is_field_based(enum bpf_core_relo_kind kind)
-{
-	switch (kind) {
-	case BPF_CORE_FIELD_BYTE_OFFSET:
-	case BPF_CORE_FIELD_BYTE_SIZE:
-	case BPF_CORE_FIELD_EXISTS:
-	case BPF_CORE_FIELD_SIGNED:
-	case BPF_CORE_FIELD_LSHIFT_U64:
-	case BPF_CORE_FIELD_RSHIFT_U64:
-		return true;
-	default:
-		return false;
-	}
-}
-
-static int check_core_relo(const struct btf *btf, const struct bpf_core_relo *rec)
-{
-#if 0
-	const char *name, *spec_str;
-
-	/*
-	 * Here is the structure defined in our eBPF program. This structure requires
-	 * relocation correction; the offset for 'name' is actually incorrect.
-	 * The offset information for this is recorded in the relocation data in the
-	 * `.BTF.ext` section, specifically in `rec->access_str_off`.
-	 *
-	 * struct filename {
-	 *    const char *uptr;
-	 *    const char *name;
-	 *    int refcnt;
-	 *    void *aname;
-	 *    const char iname[];
-	 * };
-	 *
-	 * If 'filename->name' is accessed, the resulting 'spec_str' would
-	 * be '0:1'. The initial zero-index access (&s->... is similar to &s[0]...),
-	 * and 1 represents the name field in the structure shown above. The
-	 * structure information is stored in the BTF section.
-	 */
-	spec_str = btf__name_by_offset(btf, rec->access_str_off);
-	if (str_is_empty(spec_str) || *spec_str == ':')
-		return -1;
-
-	/* Don't have a field access string, not require relocation handling. */
-	if (core_relo_is_type_based(rec->kind)) {
-		if (strcmp(spec_str, "0"))
-			return -EINVAL;
-		return 0;
-	}
-
-	// Enum values are not processed.
-	if (rec->kind == BPF_CORE_ENUMVAL_EXISTS || rec->kind == BPF_CORE_ENUMVAL_VALUE) {
-		return 0;
-	}
-
-	// Non-field type check
-	if (!core_relo_is_field_based(rec->kind))
-		return -1;
-
-	u32 id;
-	const struct btf_type *t = skip_mods_and_typedefs(btf, rec->type_id, &id);
-	if (!t)
-		return -1;
-	__s64 sz = btf__resolve_size(btf, id);
-	if (sz < 0)
-		return -1;
-
-#endif
-	return 0;
-}
-
-static void *u32_as_hash_key(__u32 x)
-{
-        return (void *)(uintptr_t)x;
-}
-
-static inline long PTR_ERR(const void *ptr)
-{
-        return (long) ptr;
-}
-
-
-static inline void * ERR_PTR(long error_)
-{
-        return (void *) error_;
-}
-
-
-extern int bpf_core_add_cands(struct bpf_core_cand *local_cand,
-                       size_t local_essent_len,
-                       const struct btf *targ_btf,
-                       const char *targ_btf_name,
-                       int targ_start_id,
-                       struct bpf_core_cand_list *cands);
-
-static struct bpf_core_cand_list *
-bpf_core_find_cands(struct ebpf_object *obj, const struct btf *local_btf, __u32 local_type_id)
-{
-        struct bpf_core_cand local_cand = {};
-        struct bpf_core_cand_list *cands;
-        const struct btf *main_btf;
-        const struct btf_type *local_t;
-        const char *local_name;
-        size_t local_essent_len;
-        int err, i;
-
-        local_cand.btf = local_btf;
-        local_cand.id = local_type_id;
-        local_t = btf__type_by_id(local_btf, local_type_id);
-        if (!local_t)
-                return ERR_PTR(-EINVAL);
-
-        local_name = btf__name_by_offset(local_btf, local_t->name_off);
-        if (str_is_empty(local_name))
-                return ERR_PTR(-EINVAL);
-        local_essent_len = bpf_core_essential_name_len(local_name);
-
-        cands = calloc(1, sizeof(*cands));
-        if (!cands)
-                return ERR_PTR(-ENOMEM);
-
-        main_btf = obj->btf_vmlinux;
-        err = bpf_core_add_cands(&local_cand, local_essent_len, main_btf, "vmlinux", 1, cands);
-        if (err)
-                goto err_out;
-
-        /* if vmlinux BTF has any candidate, don't got for module BTFs */
-        if (cands->len)
-                return cands;
-
-        return cands;
-err_out:
-        bpf_core_free_cands(cands);
-        return ERR_PTR(err);
-}
-
-static int bpf_core_resolve_relo(struct ebpf_prog *prog,
-				 const struct btf *btf,
-				 struct hashmap *cand_cache,
-				 int relo_idx, const struct bpf_core_relo *rec, struct bpf_core_relo_res *targ_res)
-{
-	struct bpf_core_spec specs_scratch[3] = {};
-	struct bpf_core_cand_list *cands = NULL;
-	const void *type_key = u32_as_hash_key(rec->type_id);
-	const struct btf_type *local_type;
-	const char *local_name;
-	int err;
-
-	// BTF type ID of the "root" (containing) entity of a relocatable
-	u32 local_id = rec->type_id;
-	local_type = btf__type_by_id(btf, local_id);
-	if (!local_type)
-		return -1;
-	local_name = btf__name_by_offset(btf, local_type->name_off);
-	if (!local_name)
-		return -1;
-	// prog 'kfunc__do_unlinkat': relo #0: target candidate search [101] 'struct filename'
-	ebpf_debug
-	    ("prog '%s': relo #%d: target candidate search [%d] '%s %s'\n",
-	     prog->name, relo_idx, local_id, btf_kind_str(local_type),
-	     local_name);
-
-	if (rec->kind != BPF_CORE_TYPE_ID_LOCAL &&
-	    !hashmap__find(cand_cache, type_key, (void **)&cands)) {
-		cands = bpf_core_find_cands(prog->obj, btf, local_id);
-		if (DF_IS_ERR(cands)) {
-			ebpf_warning("prog '%s': relo #%d: target candidate search failed for [%d] %s %s: %ld\n",
-				prog->name, relo_idx, local_id, btf_kind_str(local_type),
-				local_name, (long)cands);
-			return (long)cands;
-		}
-		err = hashmap__set(cand_cache, type_key, cands, NULL, NULL);
-		if (err) {
-			bpf_core_free_cands(cands);
-			return err;
-		}
-	}
-
-	// check_core_relo(btf, rec);
-
-	//// prog 'kfunc__do_unlinkat': relo #0: <byte_off> [101] struct filename.name (0:1 @ offset 8)
-	//ebpf_debug("prog '%s': relo #%d: %s\n", prog->name, relo_idx, spec_buf);
-
-	return bpf_core_calc_relo_insn(prog->name, rec, relo_idx, btf, cands, specs_scratch,
-				       targ_res);
-}
-
-static size_t bpf_core_hash_fn(const void *key, void *ctx)
-{
-        return (size_t)key;
-}
-
-static bool bpf_core_equal_fn(const void *k1, const void *k2, void *ctx)
-{
-        return k1 == k2;
-}
-
-// bpf_object__relocate_core
-/*
- * Clang has a built-in attribute __attribute__((preserve_access_index))
- * (equivalent to __builtin_preserve_access_index). Uses this attribute to
- * mark all the structures it needs to access. Clang generates a 'bpf_core_relo'
- * for each such access in the object ELF file.
- */
-static int obj_relocate_core(struct ebpf_prog *prog)
-{
-	const struct btf_ext_info_sec *sec;
-	const struct bpf_core_relo *rec;
-	const struct btf_ext_info *seg;
-	const char *sec_name;
-	int i, err = 0, insn_idx, sec_idx, sec_num;
-	struct sec_desc *desc = prog->sec_desc;
-	struct ebpf_object *obj = prog->obj;
-	struct bpf_core_relo_res targ_res;
-	struct bpf_insn *insn;
-
-	struct hashmap_entry *entry;
-	struct hashmap *cand_cache = NULL;
-	cand_cache = hashmap__new(bpf_core_hash_fn, bpf_core_equal_fn, NULL);
-	if (DF_IS_ERR(cand_cache)) {
-		err = PTR_ERR(cand_cache);
-		return -1;
-	}
-
-	seg = &obj->btf_ext->core_relo_info;
-	sec_num = 0;
-	// Traverse the section of BTF extended information.
-	for_each_btf_ext_sec(seg, sec) {
-		sec_idx = seg->sec_idxs[sec_num];
-		sec_num++;
-		sec_name = btf__name_by_offset(obj->btf, sec->sec_name_off);
-		if (str_is_empty(sec_name)) {
-			return -1;
-		}
-		// Traverse the CO-RE relocation information.
-		for_each_btf_ext_rec(seg, sec, i, rec) {
-			if (rec->insn_off % BPF_INSN_SZ)
-				return -1;
-			insn_idx = rec->insn_off / BPF_INSN_SZ;
-			// Verify whether this PORG contains BTF relocation information.
-			if (strcmp(sec_name, desc->name) == 0
-			    && sec_idx == desc->shndx
-			    && prog_contains_insn(prog, insn_idx)) {
-				// Adjust to the local program's instruction index.
-				insn_idx = insn_idx - prog->sec_insn_off;
-				// e.g.: sec 'fentry/do_unlinkat': found 2 CO-RE relocations,
-				// sec_idx 36 shndx 36 shndx_rel 105, insn_idx 30 prog->insns_cnt 57
-				ebpf_debug
-				    ("sec '%s': found %d CO-RE relocations, sec_idx %d shndx %ld "
-				     "shndx_rel %ld, insn_idx %d prog->insns_cnt %d\n",
-				     sec_name, sec->num_info, sec_idx,
-				     desc->shndx, desc->shndx_rel, insn_idx,
-				     prog->insns_cnt);
-
-				if (insn_idx >= prog->insns_cnt)
-					return -1;
-				insn = &prog->insns[insn_idx];
-				err = bpf_core_resolve_relo(prog, obj->btf, cand_cache, i, rec, &targ_res);
-				if (err) {
-					ebpf_warning("prog '%s': relo #%d: failed to relocate: %d\n",
-						prog->name, i, err);
-					goto out;
-				}
-				err = bpf_core_patch_insn(prog->name, insn, insn_idx, rec, i, &targ_res);
-				if (err) {
-					ebpf_warning("prog '%s': relo #%d: failed to patch insn #%u: %d\n",
-					prog->name, i, insn_idx, err);
-					goto out;
-				}
-			}
-
-		}
-	}
-
-out:
-        if (!DF_IS_ERR_OR_NULL(cand_cache)) {
-                hashmap__for_each_entry(cand_cache, entry, i) {
-                        bpf_core_free_cands(entry->value);
-                }
-                hashmap__free(cand_cache);
-        }
-	return err;
-}
-
 static int load_obj__progs(struct ebpf_object *obj)
 {
 	int i;
@@ -911,6 +562,8 @@ static int load_obj__progs(struct ebpf_object *obj)
 		 * helpers in the kernel.
 		 */
 		sanitize_prog_instructions(obj, new_prog);
+
+		// Modify eBPF instructions based on BTF relocation information.
 		obj_relocate_core(new_prog);
 
 		new_prog->prog_fd =
@@ -1004,7 +657,7 @@ static int ebpf_btf_ext_collect(struct ebpf_object *obj)
 			sec_num++;
 
 			const char *sec_name =
-			    btf__name_by_offset(obj->btf, sec->sec_name_off);
+			    btf_name_by_offset(obj->btf, sec->sec_name_off);
 			if (str_is_empty(sec_name))
 				continue;
 
@@ -1023,214 +676,6 @@ static int ebpf_btf_ext_collect(struct ebpf_object *obj)
 static inline bool is_ldimm64_insn(struct bpf_insn *insn)
 {
 	return insn->code == (BPF_LD | BPF_IMM | BPF_DW);
-}
-
-/*
- * The sym_is_extern function is used to identify symbols declared
- * as external in the current ELF object file. These symbols are not
- * defined within the current object file (their definitions are
- * expected to be found in other object files or libraries). The
- * function ensures that the symbol is undefined, has the appropriate
- * binding attribute (global or weak), and does not have a specific type.
- */
-static bool sym_is_extern(const GElf_Sym * sym)
-{
-	int bind = ELF64_ST_BIND(sym->st_info);
-	return (sym->st_shndx == SHN_UNDEF) &&
-	    (bind == STB_GLOBAL || bind == STB_WEAK) &&
-	    (ELF64_ST_TYPE(sym->st_info) == STT_NOTYPE);
-}
-
-static bool is_call_insn(const struct bpf_insn *insn)
-{
-	return insn->code == (BPF_JMP | BPF_CALL);
-}
-
-static bool sym_is_subprog(const GElf_Sym * sym)
-{
-	int bind = ELF64_ST_BIND(sym->st_info);
-	int type = ELF64_ST_TYPE(sym->st_info);
-
-	/* local function */
-	if (bind == STB_LOCAL && type == STT_SECTION)
-		return true;
-
-	/* global function */
-	return bind == STB_GLOBAL && type == STT_FUNC;
-}
-
-static int ebpf_obj__relos_collect(struct ebpf_object *obj)
-{
-	int i;
-	u32 insn_idx;
-	int progs_cnt = obj->elf_info.prog_sec_cnt;
-	/*
-	 * All sections of type `SHT_REL` are recorded here, meaning all sections
-	 * with relocations are saved. The current section is recorded as `desc->shndx`,
-	 * and the relocated section is recorded as `desc->shndx_rel`.
-	 */
-	struct sec_desc *prog_descs = obj->elf_info.prog_secs;
-
-	Elf_Scn *scn_syms;	// SHT_SYMTAB type section
-	struct bpf_insn *insns;
-	GElf_Sym sym;
-	//struct ebpf_prog *new_prog;
-
-	scn_syms = elf_getscn(obj->elf_info.elf, obj->elf_info.syms_sec->shndx);
-	if (scn_syms == NULL) {
-		ebpf_warning("elf_getscn() is failed.\n");
-		return ETR_INVAL;
-	}
-
-	for (i = 0; i < progs_cnt; i++) {
-		Elf_Scn *scn_rel;	// SHT_REL type section
-		Elf_Scn *scn;
-		Elf_Data *data_rel;	// SHT_REL type section data
-		Elf_Data *scn_data;
-		struct sec_desc *desc;
-		size_t sec_idx;
-		GElf_Shdr shdr, shdr_rel;
-		GElf_Rel rel;
-		int err, j, nrels;
-
-		desc = &prog_descs[i];
-		scn_rel = elf_getscn(obj->elf_info.elf, desc->shndx_rel);
-		if (gelf_getshdr(scn_rel, &shdr_rel) == NULL) {
-			ebpf_warning("gelf_getshdr() is failed.\n");
-			return ETR_INVAL;
-		}
-
-		data_rel = elf_getdata(scn_rel, 0);
-		if (data_rel == NULL) {
-			ebpf_warning("gelf_getshdr() is failed.\n");
-			return ETR_INVAL;
-		}
-		sec_idx = shdr_rel.sh_info;
-
-		scn = elf_getscn(obj->elf_info.elf, desc->shndx);
-		if (gelf_getshdr(scn, &shdr) == NULL) {
-			ebpf_warning("gelf_getshdr() is failed.\n");
-			return ETR_INVAL;
-		}
-		scn_data = elf_getdata(scn, 0);
-		if (scn_data == NULL) {
-			ebpf_warning("gelf_getshdr() is failed.\n");
-			return ETR_INVAL;
-		}
-		const char *relo_sec_name = elf_strptr(obj->elf_info.elf,
-						       obj->elf_info.syms_sec->
-						       strtabidx,
-						       shdr_rel.sh_name);
-		if (str_is_empty(relo_sec_name)) {
-			ebpf_warning
-			    ("elf: failed to get section name string at offset %zu\n",
-			     shdr_rel.sh_name);
-			return ETR_INVAL;
-		}
-		// e.g.: '.relfentry/do_unlinkat': collecting relocation for section(28) 'fentry/do_unlinkat'
-		ebpf_debug
-		    ("+ sec '%s': collecting relocation for section(%zu) '%s'\n",
-		     relo_sec_name, sec_idx, desc->name);
-
-		nrels = shdr_rel.sh_size / shdr_rel.sh_entsize;
-		for (j = 0; j < nrels; j++) {
-			gelf_getrel(data_rel, j, &rel);
-			// Find the symbol section corresponding to the relocated section.
-			if (find_sym_by_idx
-			    (obj->elf_info.elf, scn_syms,
-			     GELF_R_SYM(rel.r_info), &sym) != ETR_OK) {
-				ebpf_warning("Not find symbol, sym_idx:%d\n",
-					     GELF_R_SYM(rel.r_info));
-				return ETR_INVAL;
-			}
-
-			if (rel.r_offset % BPF_INSN_SZ
-			    || rel.r_offset >= scn_data->d_size) {
-				ebpf_warning
-				    (">> sec '%s': invalid offset 0x%zx for relo #%d\n",
-				     relo_sec_name, (size_t)rel.r_offset, j);
-				return ETR_INVAL;
-			}
-
-			const char *sym_name;
-			insn_idx = rel.r_offset / BPF_INSN_SZ;
-			/* relocations against static functions are recorded as
-			 * relocations against the section that contains a function;
-			 * in such case, symbol will be STT_SECTION and sym.st_name
-			 * will point to empty string (0), so fetch section name
-			 * instead
-			 */
-			if (ELF64_ST_TYPE(sym.st_info) == STT_SECTION
-			    && sym.st_name == 0) {
-				GElf_Shdr __shdr;
-				Elf_Scn *__scn =
-				    elf_getscn(obj->elf_info.elf, sym.st_shndx);
-				if (gelf_getshdr(__scn, &__shdr) == NULL) {
-					ebpf_warning
-					    ("gelf_getshdr() is failed.\n");
-					return ETR_INVAL;
-				}
-
-				sym_name =
-				    elf_strptr(obj->elf_info.elf,
-					       obj->elf_info.
-					       syms_sec->strtabidx,
-					       __shdr.sh_name);
-			} else
-				sym_name =
-				    elf_strptr(obj->elf_info.elf,
-					       obj->elf_info.
-					       syms_sec->strtabidx,
-					       sym.st_name);
-
-			sym_name = sym_name ? : "<?";
-			// e.g.: >> sec '.reluprobe/openssl_read_exit': relo #61: insn #6779 against '__protocol_filter'
-			ebpf_debug
-			    (">> sec '%s': relo #%d: insn #%u against '%s'\n",
-			     relo_sec_name, j, insn_idx, sym_name);
-
-			struct bpf_insn *insns = desc->d_buf;	// We need to add the section offset. Here, the offset is 0, so no processing is done here.
-			struct bpf_insn *insn = &insns[insn_idx];
-			if (!is_call_insn(insn) && !is_ldimm64_insn(insn)) {
-				ebpf_warning
-				    ("prog '%s': invalid relo against '%s' for insns[%d].code 0x%x\n",
-				     desc->name, sym_name, insn_idx,
-				     insn->code);
-				return ETR_INVAL;
-			}
-
-			if (sym_is_extern(&sym)) {
-				ebpf_warning
-				    ("prog '%s': extern relo not support extern for '%s'\n",
-				     desc->name, sym_name);
-				return ETR_INVAL;
-			}
-
-			/* sub-program call relocation */
-			if (is_call_insn(insn)) {
-				ebpf_warning
-				    ("prog '%s': not support call relo against '%s' at offset %zu\n",
-				     desc->name, sym_name,
-				     (size_t)sym.st_value);
-				return ETR_INVAL;
-			}
-
-			if (sym_is_subprog(&sym)) {
-				ebpf_warning
-				    ("prog '%s': not support subprog addr relo against '%s' at offset %zu+%d\n",
-				     desc->name, sym_name, (size_t)sym.st_value,
-				     insn->imm);
-				return ETR_INVAL;
-			}
-
-			/* If ".map" section, 'map->sec_idx != sym.st_shndx', redirection is
-			 * performed in relo_parse_and_apply().  */
-
-			desc->nr_reloc++;
-		}
-	}
-
-	return 0;
 }
 
 static int ebpf_obj__maps_collect(struct ebpf_object *obj)
@@ -1439,10 +884,6 @@ struct ebpf_object *ebpf_open_buffer(const void *buf, size_t buf_sz,
 	}
 
 	if (ebpf_btf_ext_collect(obj) != ETR_OK) {
-		goto failed;
-	}
-
-	if (ebpf_obj__relos_collect(obj) != ETR_OK) {
 		goto failed;
 	}
